% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collage_of_grobs.R
\name{collage_of_grobs}
\alias{collage_of_grobs}
\title{A Function to Create a Collage of Images}
\usage{
collage_of_grobs(graphic_objects, collage_file_name = "Collage\%03d.png")
}
\arguments{
\item{graphic_objects}{Vector of grobs Images to include in collage}

\item{collage_file_name}{Character Vector. Defaults to 'Collage%03d.png'}
}
\value{
Returns NULL; Saves the collage image as a side effect.
}
\description{
This function creates a collage from the given graphic object (Grobs)
}
\examples{

filenames[ clusterID == thisNumber ] \%>\%
map( ~get_raster_grob( .x )     ) -> grobs
collage_of_grobs( grobs )
}
