% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{A Function To Cluster hashes and HoGs}
\usage{
get_clusters(data, filenames, algorithm = "adaptive", N = 2)
}
\arguments{
\item{filenames}{Vector of Character Vector}

\item{algorithm}{Character Vector. Defaults to c( 'adaptive' )}

\item{N}{Numeric Vector. Number of clusters (ignored for adaptive; otherwise defaults to 2)}

\item{features_to_cluster}{Numeric Vector}
}
\value{
Returns a list of cluster IDs and cluster exemplar filenames
}
\description{
This function does these things...
}
\details{
DETAILS
}
\examples{
img_path <- '/Users/Karl/Dropbox/Projects/Video-Captioning/pipelines/box2/'
image_filenames <- dir( path=img_path )
clusters <- get_clusters( data = hogs[[2]][[1]]$hog, filenames = image_filenames )
clusters$IDs
clusters$exemplars
}
