% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_batch.R
\name{hash_batch}
\alias{hash_batch}
\title{A Function To Create Multiple Image Hashes}
\usage{
hash_batch(IMAGE_PATH = "./", cfg_hashsize = c(4:6, 8:10),
  cfg_hashmethod = c("average_hash", "dhash", "phash"),
  cfg_hashresize = c("nearest", "bilinear"))
}
\arguments{
\item{cfg_hashsize}{Numeric Vector. Defaults to c( 4:6,8:10 ). NOTE: hash size of 7 crashes R !!!}

\item{cfg_hashmethod}{Vector of Character Vector. Defaults to c( 'average_hash', 'dhash', 'phash' )}

\item{cfg_hashresize}{Vector of Character Vector. Defaults to c( 'nearest', 'bilinear' )}
}
\value{
Returns a list of hashes
}
\description{
This function does these things:
(1) Create descriptions for Hashes
(2) Get all combinations of hash size, resizing method, and hashing method
(3) Create Hashes for each image using each configuration.
}
\examples{
hashes <- hash_batch('/Users/Karl/Dropbox/Projects/Video-Captioning/pipelines/box2/')
hashes$names
hashes$values
}
