% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hog_batch.R
\name{hog_batch}
\alias{hog_batch}
\title{TITLE A Function To Create Multiple Histograms of Oriented Gradients (HoGs)}
\usage{
hog_batch(IMAGE_PATH = "./", cfg_cells = 4:9, cfg_bins = 7:15)
}
\arguments{
\item{cfg_cells}{Numeric Vector. Defaults to 4:9}

\item{cfg_bins}{Numeric Vector. Defaults to 7:15}
}
\value{
Returns a list of HoGs
}
\description{
DESCRIPTION This function does these things...
}
\details{
DETAILS

Methods include:

* this one

* that one
}
\examples{
hogs <- hog_batch( IMAGE_PATH = '/Users/Karl/Dropbox/Projects/Video-Captioning/pipelines/box2/')
hogs$names
str( hogs$values )
}
